minor refactor 2
